pyproject.toml
[project]
name = "ddm-sdk"
version = "0.1.0"
description = "DDM Python SDK"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "web3==7.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -s"
pythonpath = ["."]